리뷰 숙소 에러 처리 및 mypage/review 재생성수정 및  페이지있는버전 추가

+++ houssg/src/main/java/ssg/com/houssg/controller/ReviewController.java
@@ -35,6 +35,7 @@ import ssg.com.houssg.dto.AccommodationDto;
 import ssg.com.houssg.dto.ResponseWrapper;
 import ssg.com.houssg.dto.ReviewDto;
 import ssg.com.houssg.dto.ReviewParam;
+import ssg.com.houssg.service.ReservationService;
 import ssg.com.houssg.service.ReviewService;
 
 @RestController
@@ -48,6 +49,9 @@ public class ReviewController {
 	
 	@Autowired
 	Cloudinary cloudinary;
+	
+	@Autowired
+	ReservationService reservice;
 	
 	// 리뷰 추가
 	@PostMapping(value = "review/add", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -107,7 +111,8 @@ public class ReviewController {
 	        dto.setId(userId);
 
 	        service.addReview(dto);
-
+	        reservice.statusUpdate(reviewDto.getReservationNumber());
+	        
 	        // 리뷰 등록 성공 시
 	        return ResponseEntity.ok("리뷰 등록 성공");
 	    } catch (Exception e) {
@@ -118,6 +123,22 @@ public class ReviewController {
 
 	// my 리뷰 보기
 	@GetMapping("mypage/review")
+	public ResponseEntity<List<ReviewDto>> getMyReview(HttpServletRequest httpRequest) {
+	    System.out.println("나의 리뷰 보기");
+	    String token = getTokenFromRequest(httpRequest);
+        String userId = getUserIdFromToken(token);
+        
+	    List<ReviewDto> reviews = service.getMyReview(userId);
+
+	    if (reviews.isEmpty()) {
+	        // 리뷰가 없는 경우
+	    	return new ResponseEntity<>(new ArrayList<ReviewDto>(),HttpStatus.OK);
+	    } else {
+	        // 리뷰를 찾은 경우
+	        return ResponseEntity.ok(reviews); // 리뷰 목록 반환
+	    }
+	}
+	@GetMapping("mypage/review/page")
 	public ResponseEntity<ResponseWrapper<ReviewDto>> getMyReview(HttpServletRequest httpRequest,
 	                                                   @RequestParam int pageSize,
 	                                                   @RequestParam int page) {
@@ -131,7 +152,7 @@ public class ReviewController {
 	    String userId = getUserIdFromToken(token);
 	    ReviewParam param = new ReviewParam(userId, pageSize, page, start);
 	    // 데이터베이스 쿼리에 start와 end를 사용하여 데이터 범위를 지정
-	    List<ReviewDto> reviews = service.getMyReview(param);
+	    List<ReviewDto> reviews = service.getMyReviewPage(param);
 	    int total = service.reviewCount(param);
 
 	    ResponseWrapper<ReviewDto> responseWrapper = new ResponseWrapper<>(reviews, total);
@@ -145,8 +166,11 @@ public class ReviewController {
 	}
 	// 숙소에 관한 리뷰
 	@GetMapping("review/all/accom")
-	public ResponseEntity<List<ReviewDto>> getAllReview(@RequestParam int accomNumber) {
+	public ResponseEntity<?> getAllReview(@RequestParam Integer accomNumber) {
 	    System.out.println("숙소에 관한 리뷰 보기");
+	    if (accomNumber == 0) {
+	    	return new ResponseEntity<>("객실번호x",HttpStatus.BAD_REQUEST);
+	    }
 	     
 	    List<ReviewDto> reviews = service.getAllReview(accomNumber);
 	     
@@ -158,22 +182,57 @@ public class ReviewController {
 	        return ResponseEntity.ok(reviews); // 리뷰 목록 반환
 	    }
 	}
+	@GetMapping("review/all/accom/page")
+	public ResponseEntity<?> getAllReviewPage(@RequestParam Integer accomNumber,
+	                                          @RequestParam int pageSize,
+	                                          @RequestParam int page) {
+	    System.out.println("숙소에 관한 리뷰 보기");
+	    int start = (page - 1) * pageSize;
+	    int end = page * pageSize;
+	    
+	    int total = 0;
+	    if (accomNumber == 0) {
+	        return ResponseEntity.badRequest().body("객실번호가 유효하지 않습니다.");
+	    }
+	    
+	    ReviewParam param = new ReviewParam(accomNumber, pageSize, page, start);
+	    List<ReviewParam> reviews = service.getAllReviewPage(param);
+	    System.out.println(reviews.toString());
+	    if (reviews.isEmpty()) {
+	        // 리뷰가 없는 경우
+	    	ResponseWrapper<ReviewParam> responseWrapper = new ResponseWrapper<>(new ArrayList<>(), total);
+	        return ResponseEntity.ok(responseWrapper);
+	    } else {
+	        // 리뷰를 찾은 경우
+	    	total = service.pageTotal(param);
+	    	ResponseWrapper<ReviewParam> responseWrapper = new ResponseWrapper<>(reviews,total);
+	        return ResponseEntity.ok(responseWrapper); // 리뷰 목록 반환
+	    }
+	}
 	@PatchMapping("review/report")
-	public ResponseEntity<String> updateReview(@RequestParam int reviewNumber,@RequestParam String reportMessage) {
+	public ResponseEntity<String> updateReview(@RequestParam Integer reviewNumber,@RequestParam String reportMessage) {
 		    System.out.println("리뷰 신고하기");
+		    if (reviewNumber == 0) {
+		    	return new ResponseEntity<>("리뷰번호x",HttpStatus.BAD_REQUEST);
+		    }
+		    if (reportMessage==null) {
+		    	return new ResponseEntity<>("메세지x",HttpStatus.BAD_REQUEST);
+		    }
   
 		    int count = service.updateReview(reviewNumber,reportMessage);
 		    
 		    if (count == 0) {
-		        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("NO");
+		        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("리뷰신고x");
 		    }
 		    
-		    return ResponseEntity.ok("YES");
+		    return ResponseEntity.ok("신고완료");
 		}
 	@DeleteMapping("review")
 	public ResponseEntity<String> deleteReview(@RequestParam int reviewNumber) {
 		    System.out.println("리뷰 삭제하기");
-		    
+		    if (reviewNumber == 0) {
+		    	return new ResponseEntity<>("리뷰번호x",HttpStatus.BAD_REQUEST);
+		    }
 		    int count = service.deleteReview(reviewNumber);
 		    
 		    if (count == 0) {
@@ -200,7 +259,12 @@ public class ReviewController {
 	public ResponseEntity<String> addComment(@RequestParam int reviewNumber, @RequestParam String reviewComment) {
 	    try {
 	        System.out.println("답글 추가 시작");
-
+	        if (reviewNumber == 0) {
+		    	return new ResponseEntity<>("리뷰번호x",HttpStatus.BAD_REQUEST);
+		    }
+		    if (reviewComment==null) {
+		    	return new ResponseEntity<>("내용x",HttpStatus.BAD_REQUEST);
+		    }
 	        // 댓글을 추가하는 비즈니스 로직 수행
 	        int count = service.addComment(reviewNumber, reviewComment);
 
@@ -220,7 +284,12 @@ public class ReviewController {
 	@PatchMapping("review/comment")
 	public ResponseEntity<String> updateComment(@RequestParam int reviewNumber, @RequestParam String reviewComment) {
 	    System.out.println("답글 수정");
-	    
+	    if (reviewNumber == 0) {
+	    	return new ResponseEntity<>("리뷰번호x",HttpStatus.BAD_REQUEST);
+	    }
+	    if (reviewComment==null) {
+	    	return new ResponseEntity<>("내용x",HttpStatus.BAD_REQUEST);
+	    }
 	    int count = service.updateComment(reviewNumber, reviewComment);
 	    
 	    if (count == 1) {
@@ -230,9 +299,11 @@ public class ReviewController {
 	    }
 	}
 	@GetMapping("reservation/review")
-	public ResponseEntity<List<ReviewDto>> reservationReview(@RequestParam int reservationNumber, HttpServletRequest httpRequest) {
+	public ResponseEntity<?> reservationReview(@RequestParam Integer reservationNumber, HttpServletRequest httpRequest) {
 	    System.out.println("예약에서 나의 리뷰 보기");
-	    
+	    if (reservationNumber == null || reservationNumber <= 0) {
+            return ResponseEntity.badRequest().body("번호 x");
+        }
 	    // 토큰 및 사용자 ID를 추출하는 코드 (getTokenFromRequest 및 getUserIdFromToken 메서드 사용)
 	    String token = getTokenFromRequest(httpRequest);
 	    String userId = getUserIdFromToken(token);
